# docs/openapi.yaml
openapi: 3.1.0
info:
  title: KISY Enterprise Messenger API
  version: 0.4.0
  description: >
    REST API for the KISY enterprise messenger. Covers stages 2–4:
    authentication, access control, invitations, profile, groups, private
    chats, messages, reactions, read state, favorites, notifications, the
    CEO admin surface and the WebSocket gateway. File handling is added in a
    later stage.
servers:
  - url: /api/v1
    description: Versioned API root (behind the Nginx edge proxy)
...
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Register using an invitation token
      description: >
        Redeems a single-use invitation token (120s lifetime) and creates
        an account at the lowest clearance level, then opens the first
        session (sets auth cookies).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inviteToken, username, password]
              properties:
                inviteToken:
                  type: string
                username:
                  type: string
                  pattern: '^[A-Za-z0-9_]{3,32}$'
                password:
                  type: string
                  minLength: 12
                  maxLength: 128
      responses:
        '201':
          description: Account created, session opened
          headers:
            Set-Cookie:
              description: kisy_access and kisy_refresh HTTPOnly cookies
              schema: { type: string }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          user: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/login:
    post:
      tags: [auth]
      summary: Authenticate and open a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Authenticated, session opened
          headers:
            Set-Cookie:
              description: kisy_access and kisy_refresh HTTPOnly cookies
              schema: { type: string }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          user: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429':
          description: Rate limited or account temporarily locked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate the refresh token
      description: >
        Exchanges the current refresh cookie for a new access+refresh pair
        (refresh token rotation). Replaying a rotated token revokes the
        session (theft detection).
      security:
        - refreshCookie: []
      responses:
        '200':
          description: Tokens rotated
          headers:
            Set-Cookie:
              description: fresh kisy_access and kisy_refresh cookies
              schema: { type: string }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          accessExpiresAt: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /invites:
    post:
      tags: [invitations]
      summary: Generate an invitation token (CEO only)
      description: >
        Requires clearance level 1. Returns a single-use token valid for
        120 seconds; the plaintext token is shown only in this response.
      responses:
        '201':
          description: Invitation created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { $ref: '#/components/schemas/Invitation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /groups:
    get:
      tags: [groups]
      summary: List groups visible to the current clearance
      responses:
        '200':
          description: Visible groups (min_role_level >= actor level)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          groups:
                            type: array
                            items: { $ref: '#/components/schemas/Group' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [groups]
      summary: Create a group (CEO only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, minRoleLevel]
              properties:
                name: { type: string, minLength: 1, maxLength: 128 }
                description: { type: [string, 'null'] }
                minRoleLevel: { type: integer, minimum: 1, maximum: 10 }
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          group: { $ref: '#/components/schemas/Group' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /chats:
    get:
      tags: [chats]
      summary: List the current user's private chats
      responses:
        '200':
          description: Private chats
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          chats:
                            type: array
                            items: { $ref: '#/components/schemas/Chat' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [chats]
      summary: Open (or fetch) a private chat with another user
      description: >
        Idempotent. Enforces the initiation rule: a user may open a chat
        only with same-or-lower clearance; the existing chat is returned if
        one already exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string, format: uuid }
      responses:
        '200':
          description: Chat opened or already existed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          chat: { $ref: '#/components/schemas/Chat' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /ws:
    get:
      tags: [websocket]
      summary: Upgrade to the real-time WebSocket connection
      description: >
        Authenticated via the kisy_access cookie or an `access_token` query
        parameter. After the upgrade, the client and server exchange JSON
        frames.

        Client→Server `type`: message.send, typing.start, typing.stop,
        presence.subscribe, read.confirmation.

        Server→Client `event`: message.created, message.deleted,
        message.read, typing.started, typing.stopped, user.online,
        user.offline, error.
      parameters:
        - { name: access_token, in: query, required: false, schema: { type: string } }
      responses:
        '101': { description: Switching Protocols (WebSocket) }
        '401': { description: Authentication required or failed }
...
  /favorites:
    get:
      tags: [chats]
      summary: List favorite/pinned chats
      responses:
        '200':
          description: Favorites, pinned first
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          favorites:
                            type: array
                            items: { $ref: '#/components/schemas/Favorite' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [chats]
      summary: Add or update a favorite/pin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chatType, chatId]
              properties:
                chatType: { type: string, enum: [private, group] }
                chatId: { type: string, format: uuid }
                isPinned: { type: boolean }
                pinnedOrder: { type: [integer, 'null'] }
      responses:
        '200': { description: Stored, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [chats]
      summary: Remove a favorite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chatType, chatId]
              properties:
                chatType: { type: string, enum: [private, group] }
                chatId: { type: string, format: uuid }
      responses:
        '200': { description: Removed, content: { application/json: { schema: { $ref: '#/components/schemas/Envelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /notifications:
    get:
      tags: [notifications]
      summary: List notifications with unread count
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, default: 50, maximum: 100 } }
      responses:
        '200':
          description: Notifications
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          notifications:
                            type: array
                            items: { $ref: '#/components/schemas/Notification' }
                          unreadCount: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /admin/users:
    get:
      tags: [admin]
      summary: List accounts (CEO only)
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: offset, in: query, required: false, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          users:
                            type: array
                            items: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: kisy_access
    refreshCookie:
      type: apiKey
      in: cookie
      name: kisy_refresh
...
  schemas:
    Envelope:
      type: object
      required: [success, requestId, timestamp]
      properties:
        success: { type: boolean }
        data: { type: object }
        error: { $ref: '#/components/schemas/Error' }
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
    Error:
      type: object
      properties:
        code:
          type: string
          enum:
            - AUTH_INVALID_TOKEN
            - AUTH_INVALID_CREDENTIALS
            - AUTH_EXPIRED
            - ACCESS_DENIED
            - RESOURCE_NOT_FOUND
            - RATE_LIMITED
            - VALIDATION_FAILED
            - INTERNAL_ERROR
        message: { type: string }
        details: {}
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        roleLevel: { type: integer, minimum: 1, maximum: 10 }
        avatarUrl: { type: [string, 'null'] }
        status: { type: string, enum: [online, offline, away] }
        lastSeen: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
    Invitation:
      type: object
      properties:
        token: { type: string }
        creatorId: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time }
    Group:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: [string, 'null'] }
        avatarUrl: { type: [string, 'null'] }
        minRoleLevel: { type: integer, minimum: 1, maximum: 10 }
        createdAt: { type: string, format: date-time }
    Chat:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [private] }
        otherUserId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
...
    Favorite:
      type: object
      properties:
        chatType: { type: string, enum: [private, group] }
        chatId: { type: string, format: uuid }
        isPinned: { type: boolean }
        pinnedOrder: { type: [integer, 'null'] }
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string }
        payload: { type: object }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
...
  responses:
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Forbidden:
      description: Insufficient clearance
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    NotFound:
      description: Resource not found (or hidden from this clearance)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Conflict:
      description: Resource conflict (e.g. username taken)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
security:
  - accessCookie: []